@@ -2,7 +2,7 @@ name: example
 description: A new Flutter project.
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
