@@ -74,11 +74,15 @@ class InfoPage extends StatelessWidget {
       margin: const EdgeInsets.only(top: 24.0),
       alignment: Alignment.center,
       child: SizedBox.expand(
-        child: RaisedButton(
-          textTheme: Theme.of(context).buttonTheme.textTheme,
-          shape:
+        child: ElevatedButton(
+          style: ButtonStyle(
+            shape: MaterialStateProperty.all(
               RoundedRectangleBorder(borderRadius: BorderRadius.circular(30.0)),
-          padding: const EdgeInsets.all(8),
+            ),
+            padding: MaterialStateProperty.all(
+              const EdgeInsets.all(8),
+            ),
+          ),
           child: Text(text),
           onPressed: () => _launchURL(url),
         ),
